% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_a_per_b.R
\name{avg_a_per_b}
\alias{avg_a_per_b}
\title{Calculate the Average of 'a' per 'b'}
\usage{
avg_a_per_b(data, a, b)
}
\arguments{
\item{data}{Is the dataframe containing objects of interest.}

\item{a}{Is a column of numbers which we are primarily interested in. The parameter is named 'a', as logically, this object should be inserted into the function prior to 'b'.}

\item{b}{Is a column of numbers for which object 'a' is measure against. The parameter is named 'b', as logically, this object should come after 'a', given that we are dividing 'a' over 'b'.}
}
\value{
Returns the mean across all observations resulting from taking the average of 'a' per 'b'.
}
\description{
This function calculates the average of one object, 'a', per a second object, 'b', and then takes the mean across all observations. Object must be numerical.
}
\examples{
library(datateachr)
avg_a_per_b(apt_buildings, apt_buildings$no_of_units, apt_buildings$no_of_storeys)
avg_a_per_b(apt_buildings, apt_buildings$no_barrier_free_accessible_units, apt_buildings$no_of_units)
}
